+++ R/addItem.R
@@ -116,7 +116,7 @@ addItem <- function(dgeObj,
                    " Use a base type instead (row, col, assay, meta),",
                    " or define a new type.", sep = ""))
 
-    if (class(funArgs) == "call")
+    if ("call" %in% class(funArgs))
         funArgs <- paste(funArgs[[1]], "(",
                          paste(funArgs[2:length(funArgs)], collapse = ", "),
                          ")", sep = "")
@@ -194,9 +194,9 @@ addItems <- function(dgeObj,
                             !missing(itemList),
                             !missing(itemTypes),
                             msg = "Specify the DGEobj, itemList, and itemTypes. All are required.")
-    assertthat::assert_that(class(dgeObj)[[1]] == "DGEobj",
-                            class(itemList)[[1]] == "list",
-                            class(itemTypes)[[1]] == "list",
+    assertthat::assert_that("DGEobj" %in% class(dgeObj),
+                            "list" %in% class(itemList),
+                            "list" %in% class(itemTypes),
                             msg = "The DGEobj must be of class DGEobj, while the itemList and itemTypes must both be lists.")
     assertthat::assert_that(length(itemList) == length(itemTypes),
                             msg = "The length of the itemList must match the length of the itemTypes.")
@@ -204,7 +204,7 @@ addItems <- function(dgeObj,
                             msg = "The itemList must be a named list.")
 
     if (!missing(parents))
-        assertthat::assert_that(class(parents)[[1]] == "list",
+        assertthat::assert_that("list" %in% class(parents),
                                 length(parents) == length(itemList),
                                 msg = "The parents list must be of class 'list' and of the same length as the itemList.")