+++ R/init.R
@@ -1,13 +1,13 @@
 #' Initialize with base data (primaryAssayData, row annotations, col annotations)
 #'
-#' @param primaryAssayData               A numeric matrix or dataframe with row and
+#' @param primaryAssayData     A numeric matrix or dataframe with row and
 #'   colnames. Each column represents a sample.  Each row represents and assay.
 #'   This is typically the counts matrix in a DGE RNA-Seq experiment.
-#' @param rowData              Gene, isoform or exon level annotation. Rownames
+#' @param rowData              Gene, exon, isoform or protein level annotation. Rownames
 #'   must match the rownames in primaryAssayData
 #' @param colData              A dataframe describing the experiment design.
 #'   Rownames much match the colnames(primaryAssayData)
-#' @param level                One of "gene", "isoform", or "exon"
+#' @param level                One of "gene", "exon", "isoform" or "protein"
 #' @param customAttr           (optional) Named list of attributes
 #' @param allowShortSampleIDs  Using sequential integer rownames (even if typed
 #'   as character) is discouraged and by default will abort the DGEobj creation.
@@ -21,7 +21,7 @@
 #'
 #' @examples
 #'
-#'    dgeObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))
+#'    dgeObj   <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))
 #'    MyCounts <- dgeObj$counts
 #'    geneinfo <- dgeObj$geneData
 #'    sampinfo <- dgeObj$design
@@ -34,7 +34,6 @@
 #'                                             GeneModel = "Ensembl.R89"))
 #'
 #' @import magrittr
-#' @importFrom GenomicRanges GRanges
 #' @importFrom assertthat assert_that
 #'
 #' @export
@@ -51,11 +50,12 @@ initDGEobj <- function(primaryAssayData,
                             !missing(colData),
                             !missing(rowData),
                             !missing(level),
+                            !is.null(level),
                             msg = "Specify the primaryAssayData, colData, rowData, and level. All are required to initialize a DGEobj.")
     assertthat::assert_that(is.matrix(primaryAssayData) | is.data.frame(primaryAssayData),
                             msg = "primaryAssayData must be specified as a matrix or a data.frame.")
     assertthat::assert_that(level %in% DGEobjDef$allowedLevels,
-                            msg = paste('The specified level must be one of: ', paste(DGEobjDef$allowedLevels, collapse=", ")))
+                            msg = paste('The specified level must be one of: ', paste(DGEobjDef$allowedLevels, collapse = ", ")))
     assertthat::assert_that(!is.null(rownames(primaryAssayData)),
                             !is.null(colnames(primaryAssayData)),
                             !is.null(rownames(rowData)),
@@ -83,9 +83,9 @@ initDGEobj <- function(primaryAssayData,
     )
 
     if (!allowShortSampleIDs == TRUE) {
-        suppressWarnings(
+        suppressWarnings({
             test <- as.numeric(rownames(colData))
-        )
+        })
 
         if (all(is.na(test)) == FALSE) {
             sampleCount <- nrow(colData)
@@ -101,8 +101,9 @@ initDGEobj <- function(primaryAssayData,
 
     funArgs <- match.call()
 
-    result <- try(primaryAssayData <- as.matrix(primaryAssayData), silent = TRUE)
-    if (any(class(result) == "try-error"))
+    result <- try({primaryAssayData <- as.matrix(primaryAssayData)}, silent = TRUE)
+
+    if ("try-error" %in% class(result))
         stop("Couldn't coerce primaryAssayData to a numeric matrix!")
 
     # Initialize an empty DGEobj
@@ -120,43 +121,42 @@ initDGEobj <- function(primaryAssayData,
     # Add DGEobj version (not the package version)
     # This will only change when there is a substantive change
     # to the base DGEobj structure
-    attr(dgeObj, "DGEobjDef.version") <- "1.1"
+    attr(dgeObj, "DGEobjDef.version") <- "1.2"
 
     # Load required items
+
     # primaryAssayData (counts for RNA-Seq)
     primaryAssayName = DGEobjDef$primaryAssayNames[[level]]
     dgeObj <- addItem(dgeObj,
-                      item = primaryAssayData,
-                      itemName = paste(primaryAssayName, "_orig", sep=""),
-                      itemType = paste(primaryAssayName, "_orig", sep=""),
-                      funArgs = funArgs,
-                      parent = "",
-                      init = TRUE
-    )
+                      item     = primaryAssayData,
+                      itemName = paste(primaryAssayName, "_orig", sep = ""),
+                      itemType = paste(primaryAssayName, "_orig", sep = ""),
+                      funArgs  = funArgs,
+                      parent   = "",
+                      init     = TRUE)
 
     dgeObj <- addItem(dgeObj,
-                      item = primaryAssayData,
+                      item     = primaryAssayData,
                       itemName = DGEobjDef$primaryAssayNames[[level]],
                       itemType = DGEobjDef$primaryAssayNames[[level]],
-                      funArgs = funArgs,
-                      parent = paste(primaryAssayName, "_orig", sep=""),
-                      init = TRUE
-    )
+                      funArgs  = funArgs,
+                      parent   = paste(primaryAssayName, "_orig", sep = ""),
+                      init     = TRUE)
 
     # colData
     dgeObj <- addItem(dgeObj,
-                      item = colData,
+                      item     = colData,
                       itemName = "design_orig",
                       itemType = "design_orig",
-                      funArgs = funArgs,
-                      parent = "")
+                      funArgs  = funArgs,
+                      parent   = "")
 
     dgeObj <- addItem(dgeObj,
-                      item = colData,
+                      item     = colData,
                       itemName = "design",
                       itemType = "design",
-                      funArgs = funArgs,
-                      parent = "design_orig")
+                      funArgs  = funArgs,
+                      parent   = "design_orig")
 
     # rowData
     level    <- tolower(level)
@@ -167,40 +167,53 @@ initDGEobj <- function(primaryAssayData,
     grparent <- itemName
 
     dgeObj <- addItem(dgeObj,
-                      item = rowData,
+                      item     = rowData,
                       itemName = parent,
                       itemType = parent,
-                      funArgs = funArgs,
-                      parent = "")
+                      funArgs  = funArgs,
+                      parent   = "")
 
     dgeObj <- addItem(dgeObj,
-                      item = rowData,
+                      item     = rowData,
                       itemName = itemName,
                       itemType = itemType,
-                      funArgs = funArgs,
-                      parent = parent)
+                      funArgs  = funArgs,
+                      parent   = parent)
 
     # Annotate the level
     dgeObj %<>% setAttributes(list(level = level))
 
-    result <- try({gr <- GenomicRanges::GRanges(rowData)}, silent = TRUE)
+    if (level %in% c("exon", "gene")) {
+        genomic_ranges_success <- FALSE
 
-    if (class(result) == "try-error") {
-        warning("Couldn't build a GRanges object!")
-    } else {
-        dgeObj <- addItem(dgeObj,
-                          item = gr,
-                          itemName = "granges_orig",
-                          itemType = "granges_orig",
-                          funArgs = funArgs,
-                          parent = paste(grparent, "_orig", sep = ""))
+        if (requireNamespace("GenomicRanges", quietly = TRUE)) {
 
-        dgeObj <- addItem(dgeObj,
-                          item = gr,
-                          itemName = "granges",
-                          itemType = "granges",
-                          funArgs = funArgs,
-                          parent = grparent)
+            do.call("require", list("GenomicRanges"))
+            result <- try({
+                gr <- do.call("GRanges", list(rowData))
+            }, silent = TRUE)
+
+            if (!("try-error" %in% class(result))) {
+                dgeObj <- addItem(dgeObj,
+                                  item     = gr,
+                                  itemName = "granges_orig",
+                                  itemType = "granges_orig",
+                                  funArgs  = funArgs,
+                                  parent   = paste(grparent, "_orig", sep = ""))
+
+                dgeObj <- addItem(dgeObj,
+                                  item     = gr,
+                                  itemName = "granges",
+                                  itemType = "granges",
+                                  funArgs  = funArgs,
+                                  parent   = grparent)
+                genomic_ranges_success <- TRUE
+            }
+        }
+
+        if (!genomic_ranges_success) {
+            warning("Unable to build a GRanges object for the gene/exon data.")
+        }
     }
 
     if (!missing(customAttr))