+++ R/internal.R
@@ -17,8 +17,9 @@
              meta  = "meta",
 
              geneData    = "row",
-             isoformData = "row",
              exonData    = "row",
+             isoformData = "row",
+             proteinData = "row",
              affyData    = "row",
              granges     = "row",
 
@@ -31,28 +32,33 @@
              alignQC      = "col",
 
              counts             = "assay",
+             intensities        = "assay",
              effectiveLength    = "assay",
              AffyRMA            = "assay",
              DGEList            = "assay",
              Elist              = "assay",
              isoformFrac        = "assay",
+
              corFit             = "meta",
              topTreat           = "meta",
              geneList           = "meta",  #intended for short gene lists
-             pathway            = "meta",   #should consider format standards for genelists and pathways
+             pathway            = "meta",  #should consider format standards for genelists and pathways
              URL                = "meta",
              contrast_fit_treat = "meta",
              contrastMatrix     = "meta",
+             imputationMatrix   = "meta",
 
              #types with _orig suffix are intended to store the initialized data
-             #in its original state (i.e. before subsetting)
-             geneData_orig        = "meta",
-             isoformData_orig     = "meta",
-             exonData_orig        = "meta",
-             granges_orig         = "meta",
+             #in its original state (i.e. before subsetting, normalizing)
              counts_orig          = "meta",
-             design_orig          = "meta",
+             intensities_orig     = "meta",
              effectiveLength_orig = "meta",
+             geneData_orig        = "meta",
+             exonData_orig        = "meta",
+             isoformData_orig     = "meta",
+             proteinData_orig     = "meta",
+             granges_orig         = "meta",
+             design_orig          = "meta",
              AffyRMA_orig         = "meta",
              svobj                = "meta"
     ),
@@ -62,27 +68,30 @@
                    "counts_orig",
                    "design",
                    "design_orig",
+                   "intensities",
+                   "intensities_orig",
                    "geneData",
                    "geneData_orig",
-                   "isoformData",
-                   "isoformData_orig",
                    "exonData",
                    "exonData_orig",
+                   "isoformData",
+                   "isoformData_orig",
+                   "proteinDaa",
+                   "proteinData_orig",
                    "effectiveLength",
                    "effectiveLength_orig",
                    "granges",
                    "granges_orig",
-                   "intensities",
-                   "intensities_orig",
                    "AffyRMA",
                    "AffyRMA_orig"
     ),
 
-    allowedLevels = c("gene", "isoform", "exon", "affy"),
+    allowedLevels = c("gene", "exon", "isoform", "protein", "affy"),
 
     primaryAssayNames = c(gene    = "counts",
-                          isoform = "counts",
                           exon    = "counts",
+                          isoform = "intensities",
+                          protein = "intensities",
                           affy    = "AffyRMA")
 )