+++ NEWS.md
@@ -4,6 +4,13 @@
 
 ---
 
+### v1.1
+* made GenomicRanges package optional - used only for exon/gene levels
+* added "protein" level data, reviewed/updated isoform data for proper handling
+* added imputationMatrix as a type of metadata
+* reworked reset to resolve issues and remove unneded code
+* updated tests and documentation
+
 ### v1.0.3
 * significant documentation review/updates
 * underlying structure - added a few new fields and cleaned up unneeded methods